..why was gBattlerAbility u16

+++ include/battle.h
@@ -734,6 +734,6 @@ extern u8 gChosenActionByBattler[MAX_BATTLERS_COUNT];
 extern u8 gBattleTerrain;
 extern struct MultiBattlePokemonTx gMultiPartnerParty[3];
 extern u16 gRandomTurnNumber;
-extern u16 gBattlerAbility;
+extern u8 gBattlerAbility;
 
 #endif // GUARD_BATTLE_H
